requirements.txt
oci
paramiko
python-dotenv